Adds test for broken links in the documentation
The test check for (And fail if some are found) the following things:
- file links which don't point to a valid file (this has some overlap
with mkdocs, but also catches links which do not end in .md)
- section links which don't point to a valid section. (the ./site folder
needs to be present for this test to work, so run mkdocs build before
running this test)
- Hidden files, i.e files which are not referenced to at all. If for
some reason a file is intentionally hidden, add it to the
hidden_whitelist.txt. Only hide files for a very good reason!
There are some applications files there at the moment which need to be
completed and linked to ASAP
- Urls which point to docs.csc.fi, any links to docs.csc.fi should be
done using internal links (otherwise checking for validity is much more
difficult)