Highlight redundancy info in c/ePouta flavors
The redundancy information of the VM flavors in the previous version of
the vm-flavors-and-billing.md page were either missing or not
highlighted sufficiently. The idea is that the customer, when selecting
a flavor whose disk is not stored on Ceph for example, needs to be
well-aware of the risks of losing the VM data in case of faults.

A folder containing a set of new circle icons has been added to the
docs/img folder. All icons have been produced by myself using
libreoffice, so there are no copyright issues related with their usage.
The set of colors used should be so that they are distinguishable by
color-blind people as well, and they also distinguishable when printed
in black and white. All icons are in .svg format, thus they should not
lose quality when resized.

Main changes to the vm-flavors-and-billing.md page:
- Add an introductory "About redundancy" section
- Each table summarizing the flavors (except deprecated flavors) has
a redundancy column showing the redundancy level for power, disk, and
network

Other minor changes:
- improve table headers
- fix memory amount values

Rule adopted for reporting the memory amount in GiB:
- take the value in MB from Openstack and convert it to GiB, rounding it
to integer
- if the integer is of the form XX4 or XX6, then report XX5
- if the integer is of the form XX9 or XX1, then report XX0

Rule adopted for reporting the memory amount per core in GiB:
- divide the total memory amount in GiB by the number of cores and round
to the first decimal digit

It was noted that having the disk hosted using Ceph does not only
protect from disk failures in a single node, but it also reduces
downtime in case of a generic fault in the node used by the customer.
More precisely, if a non-disk-related fault (e.g., memory fault) occurs
in the node used by a customer, the customer might need to wait for the
fault to be fixed before being able to access his/her virtual machine.
If the disks are hosted using Ceph instead, then the virtual machine can
be spawned on an alternative node.

This commit:
- changes "Disk redundancy" to "Data redundancy"
- improves the description of the three different levels of data
redundancy, highlighting the additional risks incurred when choosing to
store data locally to a node (i.e. potentially high downtime)

Update ePouta GPU flavors information

The information for the new GPU flavors in ePouta have been added using
the former notation, i.e., specifying RAID level next to the amount
of GB of the disk, no redundancy icons, etc. The information have been
now updated with the new notation.

This commit will:
- update ePouta GPU flavors information using the redundancy icons
- fix the memory values converting them to GiB more precisely