Add instructions for converting images
Taking snapshot of virtual machine instances is a desired feature, but
the images obtained are in raw format, which quickly deplete the space
available for images. Alternative and more efficient formats do exist,
but customers do not have instructions on how to convert snapshots using
these formats, and thus saving space ultimately.

This commit will:
- add a section in the page about operations on images. This sections
describes a procedure for converting a snapshot from raw format to qcow2
and upload it to OpenStack.