Fix instructions on how to expand filesystem in volume
Currently the instructions suggest to use xfs_growfs to check the size
of the filesystem. Increasing the filesystem size is indeed the
objective of the instructions, but the command should not be used to
query the status of the filesystem. This PR attempts to clarify in which
context the xfs_growfs should be used, and when xfs_info should be used
instead .