Fix flavor memory values
Previous values were considering the memory value provided by openstack
as expressed in megabytes (MB). The value was then converted to
gibigytes (GiB).
Openstack console actually works with memory values in mebibytes (MiB).
These corrections are so that the values are then presented in GiB. In
addition, the memory amount of flavor gpu.3.1gpu has also been changed
due to an excessive amount of memory initially assigned to the flavor.