Updating CmdStan information
Updating the information on within-chain parallelization with brsm and CmdStan: Example on the page refers to using CmdStan backend with brms. Within-chain parallelization (threading) requires a project-specific installation of CmdStan, because it wants to write in the write-protected CmdStan installation folder. Instructions for CmdStan installation left out, because the installation requires a lot of memory and good enough instructions would take too much space on the page.