Adding Best Practices for Client Side Encryption
Moving this page from research.csc.fi. Making it visible as a link under Working with data contents page: https://docs.csc.fi/data/datasets/dataset-sources/, but not showing it in navigation, in order to not mess the navigation (it would go under Tips page, but that is already very long).