Separate sensitive data examples into a bullet list (#1671)
* Separate the *-separated items into a bullet list

The paragraph with an inline "bullet list" was somewhat difficult to
read and grammatically problematic. Separated the items to a proper
bullet list for easier reading.

* Soften the wording about sensitive data

The "includes" sounded a bit harsh: not _all_ exological data for
example is sensitive. Further, now it sounds less like a complete list,
so readers are perhaps more likely to give their data's sensitivity
thought beyond "is it in this list yes/no".

* Fix bullet list not rendering using mkdocs

Added the required newline before the bullet list

* Remove period after last bullet list item

---------

Co-authored-by: Joonas Somero <50655931+joonas-somero@users.noreply.github.com>