Update billing.md (#1679)
* Update billing.md

Hi,

I added the general BU policy.

* Update billing.md

fix link

* Update billing.md

fixed

* Update how-to-manage-your-project.md

I noticed that in the project closure section the sentence that was explaining that data will be deleted after 90 days was a bit hidden. Please keep the edit that you prefer the most, or do not accept the change. I was just trying to make it more notable.

* Update sd-access.md

Updated SD Desktop also

* Update billing.md

added link to Desktop BU

* Update sd-access.md

fixed link

* Update sd-access.md

* Update billing.md

* Update billing.md

* Fixed internal link

---------

Co-authored-by: spousi <31243093+spousi@users.noreply.github.com>
Co-authored-by: Joonas Somero <joonas.somero@csc.fi>