Update snakemake-puhti.md
The shell stuff is under he first link, but is likely easy to overlook, added it now explicitly. 

The folders binding is very likely also needed, so added that.

apptainer/container in the box. Not sure, we did not change that during the last update. I can now only find old docs with singularity mentioned in this context, likely has been also depricated by now. Changed to container now.