Fix example script and highlight what needs to be changed (#2158)
The example script given in the document will not work directly in Puhti, despite its description, nor will it work in Mahti. The allas_conf tool is located in a different path in both systems. The documentation does mention the correct paths, but the example script uses neither.

This commit changes the example script such that it will work in Puhti, as mentioned in the description, and instructs the user on how to make it work in Mahti.