Clarify GPU models used by gpu flavors
GPU models are written only in part and it is a bit difficult to find
them because they are immersed in a paragraph. The commit adds them as
the first bullet point in the list describing the features of each of
the gpu flavors.

The commit also adds few minor improvements to the text.