Remove warning sign from re-enabled flavors
Few flavors in ePouta were marked with a warning sign to indicate that
it was temporarily not possible to launch new virtual machines with the
flavors. This was due to scheduled maintenance. The maintenance is now
completed, so the flavors have been re-enabled. This commit is to
restore the original description of the flavor, i.e., removing the
warning sign in front of them.
Please note that this is just to indicate that launching a new virtual
machine with such flavor is technically possible again, but it does not
imply anything about the availability of resources. In other words,
launching a virtual machine with these flavors can still fail, if no
sufficient resources are available.